Removed double mcsolve
